move GetBlockID on Biscuit
This aims to reduce confusion on GetBlockID usage, mostly on lack of support
for generated facts lookup.

Exposing GetBlockID on verifier is not the best approach as the verifier
is expected to run the world and work with generated facts from rules.
Instead, it makes more sense to have it on the biscuit, as a way to
lookup facts in the biscuit blocks.
