+++ verifier.go
@@ -10,7 +10,6 @@ import (
 
 var (
 	ErrMissingSymbols = errors.New("biscuit: missing symbols")
-	ErrFactNotFound   = errors.New("biscuit: fact not found")
 )
 
 type Verifier interface {
@@ -19,7 +18,7 @@ type Verifier interface {
 	AddCaveat(caveat Caveat)
 	Verify() error
 	Query(rule Rule) (FactSet, error)
-	GetBlockID(fact Fact) (int, error)
+	Biscuit() *Biscuit
 	Reset()
 	PrintWorld() string
 }
@@ -140,30 +139,8 @@ func (v *verifier) Query(rule Rule) (FactSet, error) {
 	return result, nil
 }
 
-// GetBlockID returns the first block index containing a fact
-// starting from the authority block and then each block in order they were added.
-// Note that facts generated from rules can't be searched.
-// ErrFactNotFound is returned when no matching fact is found.
-func (v *verifier) GetBlockID(fact Fact) (int, error) {
-	// don't store symbols from searched fact in the verifier table
-	symbols := v.symbols.Clone()
-	datalogFact := fact.Predicate.convert(symbols)
-
-	for _, f := range *v.biscuit.authority.facts {
-		if f.Equal(datalogFact) {
-			return 0, nil
-		}
-	}
-
-	for i, b := range v.biscuit.blocks {
-		for _, f := range *b.facts {
-			if f.Equal(datalogFact) {
-				return i + 1, nil
-			}
-		}
-	}
-
-	return 0, ErrFactNotFound
+func (v *verifier) Biscuit() *Biscuit {
+	return v.biscuit
 }
 
 func (v *verifier) PrintWorld() string {
